:recycle: MainWindow에서 EventHandler를 분리
MainWindow는 UI와 관련된 부분만 작성하고, 각 컴포넌트끼리의 상호작용은
EventHandler에 작성했습니다.
