+++ src/main_window.py
@@ -1,7 +1,6 @@
 import sys
 from PyQt5.QtWidgets import *
 from PyQt5.QtCore import Qt
-import fitz
 
 
 from pageview_widget import PageviewWidget
@@ -10,17 +9,13 @@ from idx_widget import IdxWidget
 from scale_widget import ScaleWidget
 from note_widget import NoteWidget
 from chat_widget import ChatWidget
-from chatbot_controller import ChatbotController
+from event_handler import EventHandler
 
 
 class MainWindow(QMainWindow):
 
     def __init__(self):
         super().__init__()
-        self.pdf = None
-        self.idx = 0
-        self.max_idx = 0
-        self.file_path = ''
         self.init_ui()
         self.show()
 
@@ -29,20 +24,24 @@ class MainWindow(QMainWindow):
         self.setWindowTitle('PDF Viewer')
 
         self.pageview_widget = PageviewWidget()
-        self.pageview_widget.resized.connect(self.page_resized)
         self.note_widget = NoteWidget()
         self.chat_widget = ChatWidget()
-        self.chat_widget.chatbotRequested.connect(self.load_chatbot)
         self.preview_widget = PreviewWidget()
-        self.preview_widget.idxChanged.connect(self.idx_changed)
         self.idx_widget = IdxWidget()
-        self.idx_widget.idxChanged.connect(self.idx_changed)
         self.scale_widget = ScaleWidget()
-        self.scale_widget.scaleChanged.connect(self.scale_changed)
 
-        self.init_menu_bar()
+        widgets = {}
+        widgets['pageview'] = self.pageview_widget
+        widgets['note'] = self.note_widget
+        widgets['chat'] = self.chat_widget
+        widgets['preview'] = self.preview_widget
+        widgets['idx'] = self.idx_widget
+        widgets['scale'] = self.scale_widget
+
+        actions = self.init_menu_bar()
+        self.event_handler = EventHandler(self.window(), widgets, actions)
+
         self.init_tool_bar()
-
         self.sub_splitter = QSplitter()
         self.sub_splitter.setOrientation(Qt.Orientation.Vertical)
         self.sub_splitter.addWidget(self.pageview_widget)
@@ -63,15 +62,18 @@ class MainWindow(QMainWindow):
 
     def init_menu_bar(self):
         open_action = QAction('Open', self)
-        open_action.triggered.connect(self.open_file)
         save_action = QAction('Save', self)
-        save_action.triggered.connect(self.save_file)
+
+        actions = {}
+        actions['open'] = open_action
+        actions['save'] = save_action
 
         menubar = self.menuBar()
         menubar.setNativeMenuBar(False)
         filemenu = menubar.addMenu('File')
         filemenu.addAction(open_action)
         filemenu.addAction(save_action)
+        return actions
 
     def init_tool_bar(self):
         idx_action = QWidgetAction(self)
@@ -87,63 +89,12 @@ class MainWindow(QMainWindow):
         tool_bar.setFloatable(False)
         self.addToolBar(tool_bar)
 
-    def open_file(self):
-        file_path, _ = QFileDialog.getOpenFileName(
-            self.window(), 'Open file', '', 'PDF Files (*.pdf)')
-        self.file_path = file_path
-        self.pdf = fitz.open(file_path)
-        self.idx = 0
-        self.max_idx = self.pdf.page_count - 1
-
-        self.idx_widget.set_idx(self.idx)
-        self.idx_widget.set_max_idx(self.max_idx)
-        self.preview_widget.set_pdf(self.pdf)
-        self.pageview_widget.set_page(self.pdf[0])
-        self.note_widget.load_notes(self.pdf)
-        self.chat_widget.init_initial_ui()
-
-    def save_file(self):
-        if self.pdf:
-            self.note_widget.save_note()
-            self.pdf.save(self.pdf.name, incremental=True,
-                          encryption=fitz.PDF_ENCRYPT_KEEP)
-
     # main events
     def keyPressEvent(self, event):
-        if self.pdf:
-            if event.key() == Qt.Key_Right and self.idx < self.max_idx:
-                self.idx_changed(self.idx + 1)
-            elif event.key() == Qt.Key_Left and self.idx > 0:
-                self.idx_changed(self.idx - 1)
+        self.event_handler.key_pressed(event.key())
 
     def mousePressEvent(self, event):
-        if self.pdf and event.button() == Qt.LeftButton:
-            if self.sender() != self.note_widget:
-                self.note_widget.update_note()
-            if self.sender() != self.chat_widget.question_box:
-                self.chat_widget.question_box.clearFocus()
-
-    def idx_changed(self, idx):
-        if self.pdf:
-            self.idx = idx
-            self.pageview_widget.set_page(self.pdf[idx])
-            self.preview_widget.set_idx(idx)
-            self.idx_widget.set_idx(idx)
-            self.note_widget.set_idx(idx)
-
-    def scale_changed(self, scale, scale_policy):
-        scale = self.pageview_widget.set_scale(scale, scale_policy)
-        self.scale_widget.set_scale(scale)
-
-    def page_resized(self, scale):
-        self.scale_widget.set_scale(scale)
-
-    def load_chatbot(self):
-        if self.pdf:
-            self.chatbot_controler = ChatbotController(
-                self.file_path, self.chat_widget)
-            self.chat_widget.requested.connect(
-                self.chatbot_controler.handle_request)
+        self.event_handler.mouse_pressed(event.button(), self.sender())
 
 
 if __name__ == '__main__':
